XML Detection
It can now properly detect when a folder has only a single file, as well as if a file has an .xml extension.
HeightmapTile class is new; exists for a processed file. Logic in process for assigning neighbors as well as telling if a tile should be a neighbor or not.